+++ rollup.config.js
@@ -4,7 +4,7 @@ import nodeResolve from 'rollup-plugin-node-resolve'
 import replace from 'rollup-plugin-replace'
 import uglify from 'rollup-plugin-uglify'
 
-export default {
+const build = ({NODE_ENV, format, suffix}) => ({
   external: [
     'infestines',
     'karet',
@@ -16,6 +16,7 @@ export default {
     'ramda',
     'react'
   ],
+  input: 'src/karet.util.js',
   output: {
     globals: {
       infestines: 'I',
@@ -27,20 +28,27 @@ export default {
       'prop-types': 'PropTypes',
       ramda: 'R',
       react: 'React'
-    }
+    },
+    name: 'karet.util',
+    format,
+    file: `dist/karet.util.${suffix}`
   },
   plugins: [
-    process.env.NODE_ENV &&
-      replace({'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)}),
+    NODE_ENV && replace({'process.env.NODE_ENV': JSON.stringify(NODE_ENV)}),
     nodeResolve(),
     commonjs({
       include: 'node_modules/**',
       namedExports: {
-        'node_modules/react/index.js': ['Component', 'createElement']
+        'node_modules/react/index.js': [
+          'Component',
+          'Fragment',
+          'createElement',
+          'forwardRef'
+        ]
       }
     }),
     babel(),
-    process.env.NODE_ENV === 'production' &&
+    NODE_ENV === 'production' &&
       uglify({
         compress: {
           hoist_funs: true,
@@ -50,4 +58,11 @@ export default {
         }
       })
   ].filter(x => x)
-}
+})
+
+export default [
+  build({format: 'cjs', suffix: 'cjs.js'}),
+  build({format: 'es', suffix: 'es.js'}),
+  build({format: 'umd', suffix: 'js', NODE_ENV: 'dev'}),
+  build({format: 'umd', suffix: 'min.js', NODE_ENV: 'production'})
+]